yt_rag_production/requirements.txt
streamlit
yt-dlp
groq
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
python-dotenv